pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ytget-gui"
version = "2.4.6.6"
description = "A GUI YouTube downloader powered by yt-dlp"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Namira", email = "erfannamira@gmail.com" }
]
keywords = ["youtube-downloader", "downloader", "yt-dlp", "yt-dlp-gui", "audio-downloader", "video-downloader"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13"
]

dependencies = [
    "yt-dlp",
    "PySide6",
    "requests",
    "mutagen",
    "Pillow",
    "packaging"
]

[project.urls]
Homepage = "https://github.com/ErfanNamira/ytget-gui"
Repository = "https://github.com/ErfanNamira/ytget-gui"
Issues = "https://github.com/ErfanNamira/ytget-gui/issues"

[project.scripts]
ytget-gui = "ytget_gui.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["ytget_gui*"]
exclude = ["Images*"]

[tool.setuptools.package-data]
ytget_gui = ["icon.ico", "dialogs/*", "utils/*", "widgets/*", "workers/*"]